Modularize runtime type registration
