+++ src/core/CTypeRegistration.h
@@ -1,0 +1,28 @@
+/*
+fall-of-nouraajd c++ dark fantasy game
+Copyright (C) 2026  Andrzej Lis
+
+This program is free software: you can redistribute it and/or modify
+        it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+        but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+#pragma once
+
+namespace type_registration {
+void registerObjectTypes();
+void registerCoreTypes();
+void registerHandlerTypes();
+void registerGuiTypes();
+void registerGuiPanelTypes();
+void registerGuiWidgetTypes();
+void registerGuiAnimationTypes();
+} // namespace type_registration
