+++ src/handler/CHandlerTypeRegistration.cpp
@@ -1,0 +1,33 @@
+/*
+fall-of-nouraajd c++ dark fantasy game
+Copyright (C) 2026  Andrzej Lis
+
+This program is free software: you can redistribute it and/or modify
+        it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+        but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+#include "core/CTypeRegistration.h"
+#include "core/CTypes.h"
+#include "handler/CEventHandler.h"
+#include "handler/CObjectHandler.h"
+#include "handler/CRngHandler.h"
+
+namespace type_registration {
+void registerHandlerTypes() {
+    CTypes::register_type<CGameObject>();
+    CTypes::register_type<CGameEvent, CGameObject>();
+    CTypes::register_type<CGameEventCaused, CGameEvent, CGameObject>();
+    CTypes::register_type<CObjectHandler, CGameObject>();
+    CTypes::register_type<CEventHandler, CGameObject>();
+    CTypes::register_type<CRngHandler, CGameObject>();
+}
+} // namespace type_registration
